fix(constrain): reject trailing commas in json_object, unblock the file-size gate, route every constrainer through one pipeline flag
Three things that were left open after #1095.

1. json_object accepted a trailing comma (#1096). `[1,]` and `{"a":1,}` passed
   the mask, so a reply the caller was promised it could json.loads() did not
   parse — the one contract this mode exists to keep. After a value, `,`
   returned to the state an OPENER produces, and that state legally accepts the
   closer because an empty [] / {} is valid JSON. Two states now say "a
   value/key is mandatory here". The schema FSM was never affected: it already
   carried the equivalent after_comma guard, which is exactly why only the
   schema-less path shipped the bug. The issue reported the array spelling; the
   object one was the same hole and is fixed with it.

   Test first, and generatively: a comma is inserted before every closer in
   2000 random documents (2393 failing checks before the fix, 0 after), plus
   the minimal spellings as explicit cases so a failure names the shape. The
   empty forms [] and {} are asserted to stay legal.

2. The `File size` CI job was red on main — tools/imp-server/tool_call.cpp at
   847 code LOC against a hard threshold of 800, not allowlisted. The Gemma-4
   tool-call dialect parser (its own recursive-descent grammar, nothing to do
   with the rest of the file) moves to tool_call_gemma.cpp: 847 -> 642 and the
   gate reports 0 violations. Both entry points were already declared in
   tool_call.h, so no new header surface.

3. One `needs_constrained` flag replaces the needs_json_mode / needs_schema_mode
   pair in the decode scheduler. All three consumers only ever asked for their
   disjunction, so the split bought nothing and cost correctness: regex
   requests (#1091) were in NEITHER flag, which excluded them from the
   constrained pipeline and left two device-side fast paths relying on a
   separate per-request guard to not bypass the mask. Regex and GBNF grammars
   now take the same pipelined path as json_object / json_schema.

   MEASURED NEUTRAL, and shipped for the consistency rather than a speedup:
   eager (main) vs pipelined, same build otherwise, Qwen3-4B Q8_0, best of 3
   after a clock-ramp warmup — regex 231.9 -> 228.7 tok/s, flat grammar
   224.7 -> 224.1, state-heavy recursive JSON grammar 247.6 -> 251.0. All
   within noise, output byte-identical. The per-state mask cache warms fast
   enough that there is no host latency left to hide.

Verified: test-core 601 tests / 541 passed / 0 failed; degen_suite --only
constrained now 8/8 on Qwen3-4B-Q8_0 (json_object was the failing one on main);
degen_suite --quick 30/30; the GBNF e2e surface 9/9 and four concurrent rows
each holding their own grammar; verify-fast green (decode within the 3% gate).
File-size gate green, formatting unchanged against main on every touched file.
