+++ src/runtime/engine_scheduler.cpp
@@ -1442,7 +1442,7 @@ void Engine::decode_build_inference_state_(GPUBatch& gpu_batch,
                                            std::vector<std::shared_ptr<Request>>& valid_decode,
                                            int max_ctx, cudaStream_t dec_stream,
                                            InferenceState& state, bool& needs_logprobs,
-                                           bool& needs_json_mode, bool& needs_schema_mode) {
+                                           bool& needs_constrained) {
     state.token_ids = gpu_batch.d_token_ids;
     state.positions = gpu_batch.d_positions;
     state.n_tokens = gpu_batch.total_tokens;
@@ -1529,16 +1529,19 @@ void Engine::decode_build_inference_state_(GPUBatch& gpu_batch,
     fill_recurrent_state(*valid_decode[0], state, false, dec_stream);
 
     // Check if any request needs logprobs or constrained mode
+    // ONE flag for "this batch is driven by a host-side constraint FSM".
+    // It used to be two (json / schema) and every consumer only ever asked for
+    // their disjunction — which meant a new constrainer had to be remembered in
+    // both, and regex (#1091) was in neither, so a regex request could never
+    // reach the constrained pipeline and silently stayed on the eager path.
     needs_logprobs = false;
-    needs_json_mode = false;
-    needs_schema_mode = false;
+    needs_constrained = false;
     for (const auto& r : valid_decode) {
         if (r->logprobs)
             needs_logprobs = true;
-        if (r->json_mode && r->json_schema.empty() && r->tool_constraint_tools.empty())
-            needs_json_mode = true;
-        if (!r->json_schema.empty() || !r->tool_constraint_tools.empty())
-            needs_schema_mode = true;
+        if (r->json_mode || !r->json_schema.empty() || !r->tool_constraint_tools.empty() ||
+            !r->regex_pattern.empty() || !r->grammar.empty())
+            needs_constrained = true;
     }
 
     // Schema/JSON constraints for decode. Lazily create the per-request
@@ -1627,10 +1630,9 @@ void Engine::step_decode_forward(std::vector<std::shared_ptr<Request>>& valid_de
 
     InferenceState state;
     bool needs_logprobs = false;
-    bool needs_json_mode = false;
-    bool needs_schema_mode = false;
+    bool needs_constrained = false;
     decode_build_inference_state_(gpu_batch, valid_decode, max_ctx, dec_stream, state, needs_logprobs,
-                                  needs_json_mode, needs_schema_mode);
+                                  needs_constrained);
 
     // Per-request sampling lambda
     auto sample_per_request = [&](const Tensor& logits) -> std::vector<int32_t> {
@@ -1834,8 +1836,7 @@ void Engine::step_decode_forward(std::vector<std::shared_ptr<Request>>& valid_de
         // reading step N's tokens back. Falls through to the legacy
         // synchronous collect for every non-clean case.
         bool piped = false;
-        if (!needs_logprobs && !needs_json_mode && !needs_schema_mode &&
-            pipeline_batch_eligible_(valid_decode)) {
+        if (!needs_logprobs && !needs_constrained && pipeline_batch_eligible_(valid_decode)) {
             piped = pipeline_enter_(valid_decode, gpu_batch, graph_idx, state, logits_out,
                                     dec_stream, tokens);
         }
@@ -2020,8 +2021,8 @@ void Engine::step_decode_forward(std::vector<std::shared_ptr<Request>>& valid_de
     }
 
     // Process outputs: logprobs extraction + token distribution
-    step_decode_process_outputs(valid_decode, tokens, decode_logits_out, needs_logprobs, needs_json_mode,
-                                needs_schema_mode, dec_stream);
+    step_decode_process_outputs(valid_decode, tokens, decode_logits_out, needs_logprobs, needs_constrained,
+                                dec_stream);
 }
 
 // =====================================================================
@@ -2031,7 +2032,7 @@ void Engine::step_decode_forward(std::vector<std::shared_ptr<Request>>& valid_de
 
 void Engine::step_decode_process_outputs(std::vector<std::shared_ptr<Request>>& valid_decode,
                                          const std::vector<int32_t>& tokens, const Tensor& decode_logits_out,
-                                         bool needs_logprobs, bool needs_json_mode, bool needs_schema_mode,
+                                         bool needs_logprobs, bool needs_constrained,
                                          cudaStream_t dec_stream) {
     Tokenizer* tok = model_->tokenizer();
 
@@ -2111,7 +2112,7 @@ void Engine::step_decode_process_outputs(std::vector<std::shared_ptr<Request>>& 
         // must be allowed back in here, or bursts only ever fire once. A park
         // for a DIFFERENT request is torn down inside the launch.
         (!async_graph_runner_.is_setup() || async_parked_req_id_ >= 0) &&
-        !needs_logprobs && !needs_json_mode && !needs_schema_mode) {
+        !needs_logprobs && !needs_constrained) {
         auto& dreq = valid_decode[0];
         // forward_decode_async only implements banned_tokens + rep/freq/presence
         // penalties device-side. Any sampling feature that requires host-side
@@ -2214,15 +2215,17 @@ void Engine::step_decode_process_outputs(std::vector<std::shared_ptr<Request>>& 
         }
     }
 
-    // Constrained requests (json_mode / json_schema) can't run the conditional
-    // loop (the grammar FSM is host-side) — launch the pipelined constrained
-    // decode instead: per tick the host enqueues mask+sample AND the next
-    // forward, hiding FSM/mask latency under GPU compute. masked_sample_async
+    // Constrained requests (json_mode / json_schema / regex / GBNF grammar)
+    // can't run the conditional loop (the FSM is host-side) — launch the
+    // pipelined constrained decode instead: per tick the host enqueues
+    // mask+sample AND the next forward, hiding FSM/mask latency under GPU
+    // compute. This matters most for the two newest constrainers, whose cold
+    // masks walk the whole vocabulary. masked_sample_async
     // covers banned tokens + greedy/top-k/top-p only — penalties or any
     // host-side sampling feature stays on the eager path.
     if (decode_graph_pool_[0].graph_path_available() && valid_decode.size() == 1 && !offload_mgr_ &&
         config_.use_cuda_graphs && !async_graph_runner_.is_setup() && !cpipe_.active && !needs_logprobs &&
-        (needs_json_mode || needs_schema_mode)) {
+        needs_constrained) {
         auto& dreq = valid_decode[0];
         // rep/freq/presence penalties and think_budget ARE supported (uploaded /
         // forced per tick like the eager path) — the server defaults
